Add DoomObjectContainer<ObjType,Idx> as a replacement for the global doom object pointers states, sprnames, mobjinfo. Work-in-progress for negative indices as these data structures support mapping the index of a stored value to an IdxType (i.e. spritenum_t). Replaced the global sizes for the doom object pointers with functions that pass through to the new DoomObjectContainer instances.

+++ common/r_sprites.cpp
@@ -50,7 +50,7 @@ int maxframe;
 void R_CacheSprite(spritedef_t *sprite)
 {
 	DPrintf ("cache sprite %s\n",
-		sprite - sprites < ::num_spritenum_t_types ? sprnames[sprite - sprites] : "");
+		sprite - sprites < ::num_spritenum_t_types() ? sprnames[(int)(sprite - sprites)] : "");
 	for (int i = 0; i < sprite->numframes; i++)
 	{
 		for (int r = 0; r < 16; r++)
